+++ Rebus.SqlServer.Tests/DisposableCallback.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Rebus.SqlServer.Tests
+{
+    class DisposableCallback : IDisposable
+    {
+        readonly Action _disposeAction;
+
+        public DisposableCallback(Action disposeAction) => _disposeAction = disposeAction;
+
+        public void Dispose() => _disposeAction();
+    }
+}